+++ src/kiara/operations/calculate_hash.py
@@ -91,7 +91,7 @@ class CalculateHashOperationType(OperationType):
     ) -> typing.Dict[str, Operation]:
 
         result = {}
-        for op_config in self.operation_configs.values():
+        for op_config in self.operations.values():
             if op_config.module_config["value_type"] == value_type:
                 result[op_config.module_config["hash_type"]] = op_config