+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - refactoring of operation type input names -- this replaces most instances where the input name was 'value_item' (or similar). I've decided that using the value type as input name increases usability of those operations more than the costs associated with having different input names for operations of the same type:
   - pretty_print: 'value_item' -> type name of the value to pretty print
   - extract_metadata: 'value_item' -> type name of the value to extract metadata from
+  - import: 'source' input -> input file type, 'value_item' output -> target file type
 
 ## Version 0.2.1